+++ src/C4GameVersion.h
@@ -19,6 +19,8 @@
 #include "C4Version.h"
 #include "C4InputValidation.h"
 
+#include <cinttypes>
+
 struct C4GameVersion
 {
 	ValidatedStdStrBuf<C4InVal::VAL_NameAllowEmpty> sEngineName; // status only - not used for comparison
@@ -37,7 +39,7 @@ struct C4GameVersion
 
 	StdStrBuf GetString() const
 	{
-		return FormatString("%s %d.%d.%d.%d [%d]", sEngineName.getData(), (int)iVer[0], (int)iVer[1], (int)iVer[2], (int)iVer[3], (int)iBuild);
+		return FormatString("%s %" PRId32 ".%" PRId32 ".%" PRId32 ".%" PRId32 " [%" PRId32 "]", sEngineName.getData(), iVer[0], iVer[1], iVer[2], iVer[3], iBuild);
 	}
 
 	bool operator==(const C4GameVersion &rCmp) const
