+++ src/C4Wrappers.cpp
@@ -34,7 +34,7 @@ int32_t PixCol2MatOld(uint8_t pixc)
 
 int32_t PixCol2MatOld2(uint8_t pixc)
 {
-	int32_t iMat = ((int32_t)(pixc & 0x7f)) - 1;
+	int32_t iMat = (pixc & 0x7f) - 1;
 	// if above MVehic, don't forget additional vehicle-colors
 	if (iMat <= MVehic) return iMat;
 	// equals middle vehicle-color
@@ -52,22 +52,22 @@ int32_t PixCol2MatOld2(uint8_t pixc)
 
 void GameMsgObject(const char *szText, C4Object *pTarget, int32_t iFCol)
 {
-	Game.Messages.New(C4GM_Target, szText, pTarget, NO_OWNER, 0, 0, (uint8_t)iFCol);
+	Game.Messages.New(C4GM_Target, szText, pTarget, NO_OWNER, 0, 0, static_cast<uint8_t>(iFCol));
 }
 
 void GameMsgObjectPlayer(const char *szText, C4Object *pTarget, int32_t iPlayer, int32_t iFCol)
 {
-	Game.Messages.New(C4GM_TargetPlayer, szText, pTarget, iPlayer, 0, 0, (uint8_t)iFCol);
+	Game.Messages.New(C4GM_TargetPlayer, szText, pTarget, iPlayer, 0, 0, static_cast<uint8_t>(iFCol));
 }
 
 void GameMsgGlobal(const char *szText, int32_t iFCol)
 {
-	Game.Messages.New(C4GM_Global, szText, nullptr, ANY_OWNER, 0, 0, (uint8_t)iFCol);
+	Game.Messages.New(C4GM_Global, szText, nullptr, ANY_OWNER, 0, 0, static_cast<uint8_t>(iFCol));
 }
 
 void GameMsgPlayer(const char *szText, int32_t iPlayer, int32_t iFCol)
 {
-	Game.Messages.New(C4GM_GlobalPlayer, szText, nullptr, iPlayer, 0, 0, (uint8_t)iFCol);
+	Game.Messages.New(C4GM_GlobalPlayer, szText, nullptr, iPlayer, 0, 0, static_cast<uint8_t>(iFCol));
 }
 
 void GameMsgObjectDw(const char *szText, C4Object *pTarget, uint32_t dwClr)
